refactor(NetworksManagementView): move formatRpcUrlForDisplay to utils and clean up imports
