+++ app/components/Views/NetworksManagement/NetworksManagementView.tsx
@@ -13,8 +13,7 @@ import {
 
 import { strings } from '../../../../locales/i18n';
 import Routes from '../../../constants/navigation/Routes';
-import hideKeyFromUrl from '../../../util/hideKeyFromUrl';
-import hideProtocolFromUrl from '../../../util/hideProtocolFromUrl';
+import { formatRpcUrlForDisplay } from './NetworksManagementView.utils';
 import { useAddPopularNetwork } from '../../hooks/useAddPopularNetwork';
 import { PopularList } from '../../../util/networks/customNetworks';
 
@@ -36,24 +35,6 @@ import { SECTION_KEYS } from './NetworksManagementView.constants';
 import { NetworkManagementItem } from './NetworksManagementView.types';
 
 import AdditionalNetworkItem from './components/AdditionalNetworkItem';
-
-/** Same as NetworkSelector: short host + path, no protocol or API key. Truncated to fit in one line without changing shared cell layout. */
-const MAX_RPC_DISPLAY_LENGTH = 36;
-
-export const formatRpcUrlForDisplay = (url: string): string => {
-  const withoutKey = hideKeyFromUrl(url);
-  const withoutProtocol = hideProtocolFromUrl(withoutKey);
-  let result: string;
-  if (withoutProtocol?.startsWith('http')) {
-    result = withoutProtocol.replace(/^https?:\/\//, '');
-  } else {
-    result = withoutProtocol ?? url;
-  }
-  if (result.length > MAX_RPC_DISPLAY_LENGTH) {
-    return `${result.slice(0, MAX_RPC_DISPLAY_LENGTH - 1)}…`;
-  }
-  return result;
-};
 
 const NetworksManagementView = () => {
   const navigation = useNavigation();
